+++ server/World/Data/Models/EventModel.cs
@@ -7,12 +7,12 @@ namespace World.Data.Models
 {
     public class EventModel
     {
-        public uint EventID { get; set; }
+        public long EventID { get; set; }
         public long PlayerID { get; set; }
-        public uint EventType { get; set; }
+        public short EventType { get; set; }
         public int EventValue { get; set; }
-        public uint TileX { get; set; }
-        public uint TileY { get; set; }
+        public int TileX { get; set; }
+        public int TileY { get; set; }
         public DateTime TimeStamp { get; set; }
     }
 }
