+++ server/World/Data/Models/PathTileModel.cs
@@ -4,10 +4,10 @@ namespace World.Data.Models
 {
     public class PathTileModel
     {
-        public uint PathID { get; set; }
-        public uint TileID { get; set; }
-        public uint TileX { get; set; }
-        public uint TileY { get; set; }
+        public long PathID { get; set; }
+        public long TileID { get; set; }
+        public int TileX { get; set; }
+        public int TileY { get; set; }
         // TimeStamp since start of day
         public float TimeStamp { get; set; }
         public PathModel Path { get; set; }
